Создание проекта для WPF. Вынесение базовой логики.

+++ Disciples.Engine/GameInfo.cs
@@ -11,9 +11,9 @@
 
         //public static int Height { get; } = OriginalHeight;
 
-        public static int Width { get; } = 1440;
+        public static double Width { get; } = 1440;
 
-        public static int Height { get; } = 1080;
+        public static double Height { get; } = 1080;
 
 
         public static double Scale { get; } = (double) Height / OriginalHeight;
